requirements.txt
beautifulsoup4
polars-lts-cpu==0.20.8
prettytable
simple-term-menu
